server/requirements.rpi.txt
appdirs==1.4.4
apprise==1.0.0 
asgiref==3.5.2
attrs==22.1.0
audioread==2.1.9
birdnetlib==0.0.17
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
darker==1.5.0
decorator==5.1.1
dj-database-url==1.0.0
Django==3.2.15
django-extensions==3.2.0
django-model-utils==4.2.0
djangorestframework==3.14.0
flickr-api==0.7.7
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
joblib==1.1.0
librosa==0.9.2
llvmlite==0.39.0
mock==4.0.3
mypy-extensions==0.4.3
numba==0.56.0
numpy==1.19.5
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pooch==1.6.0
py==1.11.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
resampy==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
SoundFile==0.10.3.post1
sqlparse==0.4.2
tflite-runtime @ file:///usr/src/app/tflite_runtime-2.6.0-cp39-cp39-linux_aarch64.whl
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.11
watchdog==2.1.9
zipp==3.8.1